+++ .gitignore
@@ -2,3 +2,4 @@
 *.jl.*.cov
 *.jl.mem
 .DS_Store
+Manifest.toml
