+++ Project.toml
@@ -47,4 +47,5 @@ NFFT = "0.5"
 Plots = "1.4,1.5"
 Reexport = "0.2"
 SpecialFunctions = "0.10"
+Wavelets = "0.9"
 julia = "1.3, 1.4, 1.5"
