+++ src/mri/image_geom_mri.jl
@@ -11,7 +11,6 @@ Date Update:  October 7th 2019
 
 export image_geom_mri
 
-using Test: @test, @inferred
 #using MIRT: image_geom, MIRT_image_geom
 
 
@@ -24,15 +23,4 @@ image pixel indices go from -N/2 to N/2-1.
 """
 function image_geom_mri( ; kwargs...)
     return image_geom( ; offset_x = .5, offset_y = .5, offset_z = .5, kwargs...)
-end
-
-
-"""
-    image_geom_mri(:test)
-self test
-"""
-function image_geom_mri(test::Symbol)
-    test != :test && throw(DomainError(test, "Not valid"))
-    @test (@inferred image_geom_mri( ; nx = 64, dx = 2)) isa MIRT_image_geom
-    true
 end
