+++ src/mri/kspace-spiral.jl
@@ -9,7 +9,7 @@ that he got from Brad Sutton who got them from Doug Noll...
 export mri_kspace_spiral
 
 using Interpolations
-using Plots; default(markerstrokecolor=:auto, markersize=1, markershape=:circle)
+
 
 """
     kspace, omega, gradxy = mri_kspace_spiral( [options] )
@@ -269,39 +269,4 @@ function genspi(D, N ;
 	sy = imag(s)
 
 	return Gx, Gy, kx, ky, sx, sy, gts
-end
-
-
-"""
-    mri_kspace_spiral(:test)
-self test
-"""
-function mri_kspace_spiral(test::Symbol ; warn::Bool=false, show::Bool=false)
-	N = 64
-	test != :test && throw(DomainError(test, "Not valid"))
-
-	k0, o0, g0 = mri_kspace_spiral() # default 22,-1
-	for fov in (20,21)
-		mri_kspace_spiral( ; fov=fov, Nt=-1, warn_nk=warn)
-	end
-	k5l,_,g5l = mri_kspace_spiral(nl = 5) # interleaves
-
-	plot(xlabel="kx", ylabel="ky", aspect_ratio=1)
-	p1 = scatter!(k0[:,1], k0[:,2], label = "1-shot spiral")
-
-	p2 = plot()
-	scatter!(g0[:,1], label="gx")
-	scatter!(g0[:,2], label="gy")
-
-	p4 = plot(g5l[:,1,:], label="")
-	plot!(g5l[:,2,:], label="")
-
-	plot(xlabel="kx", ylabel="ky", aspect_ratio=1, title="5-shot spiral")
-	for ii=1:5
-		scatter!(k5l[:,1,ii], k5l[:,2,ii], label="")
-	end
-	plot(p1, p2, current(), p4)
-	show && prompt()
-
-    return true
 end
