+++ src/mri/kspace.jl
@@ -6,9 +6,6 @@ kspace sampling patterns
 
 export ir_mri_kspace_ga_radial
 
-using Plots: plot, plot!, scatter, gui
-using Test: @test, @inferred
-
 
 """
 `kspace = ir_mri_kspace_ga_radial(; Nro=?, Nspoke=?, ...)`
@@ -24,7 +21,6 @@ option
 - `shift::Real`		shift due to gradient delays, default 0
 		radial sample locations are `ir * delta_ro`
 		where `ir = [-(Nro/2 - 1):1:Nro/2] + shift`
-- `show::Bool`		plot k-space locations, default `false`
 
 out
 - `kspace`	`[Nro Nspoke 2]` (Float32)
@@ -35,15 +31,14 @@ out
 2015-07 Mai Le, original Matlab version
 2015-07-04 Jeff Fessler, minor changes to Matlab version
 """
-function ir_mri_kspace_ga_radial(;
-		Nro::Int = 256,
-		Nspoke::Int = 1,
-		delta_ro::Real = 1/Nro,
-		shift::Real = 0,
-		angle::Real = pi*(sqrt(5)-1)/2, # golden angle [radians]
-		start::Real = pi/2,
-		show::Bool = false,
-	)
+function ir_mri_kspace_ga_radial( ;
+	Nro::Int = 256,
+	Nspoke::Int = 1,
+	delta_ro::Real = 1/Nro,
+	shift::Real = 0,
+	angle::Real = pi*(sqrt(5)-1)/2, # golden angle [radians]
+	start::Real = pi/2,
+)
 
 	isodd(Nro) && throw(ArgumentError("Number along read out not even!"))
 
@@ -62,28 +57,5 @@ function ir_mri_kspace_ga_radial(;
 	kspace[:,:,2] .= ky
 #	kspace = cat(dims=3, kx, ky) # [Nro Nspoke 2] fails @inferred
 
-	if show
-		scatter(kx, ky, label="", aspect_ratio=1)
-		kmax = delta_ro * Nro / 2
-		plot!(xlim = [-1,1]*kmax)
-		plot!(ylim = [-1,1]*kmax)
-		plot!(xtick = (-1:1) * kmax, xlabel = "k_x")
-		plot!(ytick = (-1:1) * kmax, ylabel = "k_y")
-		gui()
-	end
-
 	return kspace
-end
-
-
-"""
-`ir_mri_kspace_ga_radial`
-self test
-"""
-function ir_mri_kspace_ga_radial(test::Symbol)
-	test != :test && throw(":test")
-	k = @inferred ir_mri_kspace_ga_radial(Nro=30, Nspoke=13, start=pi,
-		show=true, shift=-0.5, delta_ro=1)
-	@test size(k) == (30, 13, 2)
-	true
 end
