--- src/nufft/z-test.jl
@@ -1,6 +1,0 @@
-# nufft/z-test.jl
-
-using Test: @test
-
-@test dtft(:test)
-@test nufft(:test)
