+++ src/plot/z-list.jl
@@ -1,4 +1,3 @@
 # plot/z-list.jl
 
 include("jim.jl")
-include("setup.jl")
