+++ test/io/fld-read.jl
@@ -1,0 +1,8 @@
+# fld-read.jl
+
+using MIRT: fld_header, fld_read
+using Test: @test
+
+
+# todo: test read slices when that feature is added
+# fld_write ...
