+++ test/io/shows.jl
@@ -1,0 +1,7 @@
+# shows.jl
+
+using MIRT: @shows
+using Test: @test
+
+var = ones(3,4)
+@test (@shows var) isa Nothing
