+++ test/nufft/z-test.jl
@@ -1,0 +1,16 @@
+# nufft/z-test.jl
+
+using Test: @testset
+
+# nufft/z-list.jl
+
+list = [
+"dtft.jl"
+"nufft.jl"
+]
+
+for file in list
+	@testset "$file" begin
+		include(file)
+	end
+end
